eliminated top level event package, content is only used on single places
